+++ R/class_measures.R
@@ -1,5 +1,5 @@
 make_node_measure <- function(out, .data) {
-  if(is_labelled(.data)) names(out) <- node_names(.data)
+  if(is_labelled(.data)) names(out) <- node_labels(.data)
   class(out) <- c("node_measure", class(out))
   attr(out, "mode") <- node_is_mode(.data)
   out