+++ R/class_motifs.R
@@ -2,7 +2,7 @@
 make_node_motif <- function(out, .data) {
   class(out) <- c("node_motif", class(out))
   if(is_twomode(.data)) attr(out, "mode") <- node_is_mode(.data)
-  if(is_labelled(.data)) attr(out, "dimnames")[[1]] <- node_names(.data)
+  if(is_labelled(.data)) attr(out, "dimnames")[[1]] <- node_labels(.data)
   out
 }